Refactor Main.scala of the perf runner completely & remove dependency on user provided JWT's
