+++ ledger-service/http-json-perf/BUILD.bazel
@@ -57,7 +57,6 @@ perf_runtime_deps = {
         deps = [
             "//language-support/scala/bindings-akka",
             "//ledger-api/rs-grpc-bridge",
-            "//ledger-service/fetch-contracts",
             "@maven//:io_gatling_gatling_netty_util",
             "@maven//:io_netty_netty_common",
             "@maven//:io_netty_netty_transport",
